fix: priority input no longer jumps around on keystroke
Removed aggressive clamping during onChange — now accepts raw input
while typing. Clamping applied only on save (1-100 for indexers,
1-10 for download clients).

+++ crates/stackarr-core/src/models/discover.rs
@@ -31,7 +31,7 @@ pub enum DiscoverSliderType {
 #[derive(Debug, Clone, Serialize, Deserialize, FromRow)]
 #[serde(rename_all = "camelCase")]
 pub struct DiscoverSlider {
-    pub id: i64,
+    pub id: i32,
     pub slider_type: DiscoverSliderType,
     pub display_order: i32,
     pub is_built_in: bool,
@@ -66,5 +66,5 @@ pub struct UpdateDiscoverSliderInput {
 #[serde(rename_all = "camelCase")]
 pub struct ReorderSlidersInput {
     /// Ordered list of slider IDs in the desired display order.
-    pub slider_ids: Vec<i64>,
+    pub slider_ids: Vec<i32>,
 }
